requirements.txt
asgiref==3.8.1
click==8.1.8
Django==4.2.23
djangorestframework==3.16.0
drf-yasg==1.21.10
h11==0.16.0
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
uvicorn==0.34.3
